requirements.txt
numpy
pandas
Pillow
tensorflow
scikit-learn
matplotlib
